fix: add test fixture to avoid RpcEnv error

+++ src/test/scala/com/jcdecaux/setl/storage/connector/ExcelConnectorSuite.scala
@@ -6,14 +6,36 @@ import java.sql.{Date, Timestamp}
 import com.jcdecaux.setl.config.{Conf, Properties}
 import com.jcdecaux.setl.storage.SparkRepositorySuite
 import com.jcdecaux.setl.{SparkSessionBuilder, TestObject, TestObject2}
+import org.apache.commons.lang.SystemUtils
 import org.apache.log4j.{Logger, SimpleLayout, WriterAppender}
 import org.apache.spark.SparkConf
 import org.apache.spark.sql.types._
 import org.apache.spark.sql.{Dataset, SaveMode, SparkSession}
+import org.scalatest.Outcome
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 
 class ExcelConnectorSuite extends AnyFunSuite with Matchers {
+
+  override def withFixture(test: NoArgTest): Outcome = {
+    // Shared setup (run at beginning of each test)
+    SparkSession.getActiveSession match {
+      case Some(ss) => ss.stop()
+      case _ =>
+    }
+    SparkSession.clearActiveSession()
+    SparkSession.clearDefaultSession()
+    try test()
+    finally {
+      // Shared cleanup (run at end of each test)
+      SparkSession.getActiveSession match {
+        case Some(ss) => ss.stop()
+        case _ =>
+      }
+      SparkSession.clearActiveSession()
+      SparkSession.clearDefaultSession()
+    }
+  }
 
   import SparkRepositorySuite.deleteRecursively
 
@@ -190,7 +212,10 @@ class ExcelConnectorSuite extends AnyFunSuite with Matchers {
         .show()
     )
 
-    deleteRecursively(new File("src/test/resources/test_excel_sheet.xlsx"))
+    if (!SystemUtils.IS_OS_WINDOWS) {
+      // the previous assertThrows will cause the file be locked in windows. Skip the delete if OS is windows
+      deleteRecursively(new File("src/test/resources/test_excel_sheet.xlsx"))
+    }
   }
 
   test("ExcelConnector with multiple sheets") {
@@ -292,6 +317,8 @@ class ExcelConnectorSuite extends AnyFunSuite with Matchers {
     // testSheet2 is gone
     assertThrows[IllegalArgumentException](connector2.read().show())
 
-    deleteRecursively(new File(thisPath))
-  }
+    if (!SystemUtils.IS_OS_WINDOWS) {
+      // the previous assertThrows will cause the file be locked in windows. Skip the delete if OS is windows
+      deleteRecursively(new File("src/test/resources/test_excel_sheet.xlsx"))
+    }  }
 }
